+++ frontend/src/app/(components)/Header/Logo.tsx
@@ -1,6 +1,6 @@
+import LogoImage from "@/../public/assets/img/logo.png"
 import Image from "next/image"
 import { twMerge } from "tailwind-merge"
-import LogoImage from "../../public/assets/img/logo.png"
 
 type LogoProps = {
   className?: string
