+++ frontend/src/app/(components)/Header/index.tsx
@@ -1,0 +1,11 @@
+import Logo from "@/app/(components)/Header/Logo"
+import NavBar from "@/app/(components)/Header/NavBar"
+
+export default function Header() {
+  return (
+    <header className="m-14 flex w-full max-w-7xl flex-col gap-14">
+      <Logo />
+      <NavBar />
+    </header>
+  )
+}
